pyproject.toml
[project]
name = "bygeon"
version = "0.0.9"
authors = [{ name = "dummyx", email = "pygeon@gmx.com" }]
description = "A simple tool for syncing messages between multiple IMs"
readme = "README.md"
license = { text = "AGPL-3.0-only" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: Unix",
]

dependencies = [
    "requests>=2.28.1",
    "tomli>=2.0.1",
    "websocket_client>=1.4.2",
    "PyPika>=0.48.9",
    "orjson>=3.8.2",
    "typing-extensions>=4.3.0",
]

[project.scripts]
bygeon = "bygeon.main:main"

[project.urls]
"Homepage" = "https://github.com/dummyx/bygeon"
"Bug Tracker" = "https://github.com/dummyx/bygeon/issues"
